requirements.txt
alembic==1.6.5
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
greenlet==1.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jwcrypto==1.0
Mako==1.1.4
MarkupSafe==2.0.1
oauth2client==4.1.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.5.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jwt==3.3.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.20
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
